requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
pydantic
pydantic-settings
openai
requests
pypdf
pdfplumber
pymupdf
numpy
scikit-learn
sentence-transformers
streamlit
streamlit-chat
python-multipart
aiofiles
httpx
tenacity
langchain
langchain-openai
